drivelm-analysis/requirements.txt
# Core Python Dependencies
numpy>=1.21.0
pandas>=1.3.0
pathlib
typing-extensions>=4.0.0
dataclasses

# Data Processing and Analysis
scipy>=1.7.0
scikit-learn>=1.0.0
json5>=0.9.6

# Text Processing and NLP
nltk>=3.7
regex>=2022.3.15

# Visualization - Static Charts
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=8.3.0

# Visualization - Interactive Charts
plotly>=5.3.0
plotly-express>=0.4.1
kaleido>=0.2.1

# Web Application Framework
streamlit>=1.28.0

# RAG System Dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
torch>=1.9.0
transformers>=4.21.0
tokenizers>=0.13.0

# LangChain for RAG Pipeline
langchain>=0.1.0
langchain-community>=0.0.10

# Google Gemini Integration
google-generativeai>=0.3.0

# Progress and Logging
tqdm>=4.62.0
colorlog>=6.6.0

# Optional: Advanced Features
opencv-python-headless>=4.5.0
numba>=0.56.0
joblib>=1.1.0

# Development and Testing (Optional)
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Optional: GPU Support (uncomment if GPU available)
# faiss-gpu>=1.7.4  # Replace faiss-cpu with this for GPU acceleration
# torch-audio>=0.9.0  # Additional PyTorch audio support